// apps/admin/src/domains/ReportDailySale/types/DailyReportRowModal.ts
export type DailyReportRowModal = {
  orderItemId: string
  cargoId: number
  categoryId: number
  categoryName: string
  commissionAmount: number
  commissionPrice: string
  customerAddress: string
  customerCellphone: string
  customerId: number
  customerName: string
  dueDeliveryDate: Date
  duePickupDate: Date
  orderDate: Date
  originalCategoryName: string
  preparationDays: string
  productDiscount: number
  productFinalPrice: number
  productHSIN: string
  productId: number
  productName: string
  productOriginalPrice: number
  productQty: number
  productTotalPrice: number
  relatedCommissionId: number
  shippingCompanyId: string
  shippingCompanyName: string
  shippingProviderId: number
  shippingProviderName: number
  shippmentPrice: number
  shoppingCartId: number
  shoppingCartItemId: number
  stateCode: number
  stateName: string
  storeId: number
  storeName: string
  totalPayablePrice: number
  vatPercent: number
  vatPrice: number
  vendorAddress: string
  vendorId: number
  vendorName: string
  vendorShippingShare: number
  isActive: boolean
}